Define main layout

+++ AICalories/Interfaces/IMealItem.cs
@@ -10,11 +10,11 @@ namespace AICalories.Interfaces
         DateTime Date { get; set; }
         string Time { get; set; }
         string ImagePath { get; set; }
-        int Weight { get; set; }
-        int Calories { get; set; }
-        int Proteins { get; set; }
-        int Fats { get; set; }
-        int Carbohydrates { get; set; }
+        double Weight { get; set; }
+        double Calories { get; set; }
+        double Proteins { get; set; }
+        double Fats { get; set; }
+        double Carbohydrates { get; set; }
         string TotalResultJSON { get; set; }
         List<IngredientItem> Ingredients { get; set; }
     }
